+++ src/Workloads/Stacks/DotNet/DotNetSourceProject.cs
@@ -18,6 +18,10 @@ internal sealed class DotNetSourceProject(WorkingDirectory workingDirectory, str
 
     public string ProjectFilePath { get; } = projectFilePath ?? throw new ArgumentNullException(nameof(projectFilePath));
 
+    public override string? Language => Path.GetExtension(ProjectFilePath).Equals(".fsproj", StringComparison.OrdinalIgnoreCase)
+        ? "F#"
+        : "C#";
+
     public override async Task PrepareForHostRunAsync(FunctionsProjectHostRunContext context, CancellationToken cancellationToken)
     {
         ArgumentNullException.ThrowIfNull(context);